Move `server.js` entrypoint to `bin` directory
